Fix types and linting issues
