+++ client/landing/stepper/declarative-flow/internals/types.ts
@@ -46,4 +46,4 @@ export type StepProps = {
 
 export type Step = React.FC< StepProps >;
 
-export type ProvidedDependencies = Record< string, { slug: string } | boolean | string >;
+export type ProvidedDependencies = Record< string, unknown >;
